Initialize monorepo, create some packages, test usage
