+++ packages/cli/README.md
@@ -1,0 +1,11 @@
+# `@ultras/cli`
+
+> The cli tool for ultras internal projects
+
+## Usage
+
+```
+npm -g i @ultras/cli
+
+// cli --help
+```
