+++ packages/cli/__tests__/cli.test.js
@@ -1,0 +1,8 @@
+'use strict';
+
+const cli = require('../lib/cli');
+
+describe('cli cli', () => {
+  // const argv = cli(cwd).parse(['args']);
+  it('needs tests');
+});
