+++ packages/cli/bin/cli
@@ -1,0 +1,6 @@
+#!/usr/bin/env node
+
+'use strict';
+
+// eslint-disable-next-line no-unused-expressions
+require('../lib/cli')().parse(process.argv.slice(2));
