+++ packages/cli/index.js
@@ -1,0 +1,3 @@
+const { multiply } = require("@ultras/shared");
+
+console.log(multiply(1, 2));
