--- shared/cli/package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "yargs": "^17.1.1"
-  }
-}
